# docs/openapi/openapi.yaml
components:
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  contact: {}
  description: Deterministic, highly resilient backend infrastructure for the Lemni Payment-as-a-Service (PaaS) engine.
  title: LEMNI PaaS Core Engine API
  version: 1.0.0
openapi: 3.0.0
paths:
...
  /test-api-key:
    get:
      description: >-
        Verifies that the provided API key in the Authorization header is valid and authorized. Use this to test your
        API key configuration.
      operationId: AppController_testApiKey
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  status:
                    example: authorized
                    type: string
                type: object
          description: API key is valid and authorized
        '401':
          content:
            application/json:
              schema:
                properties:
                  error:
                    example: Unauthorized
                    type: string
                type: object
          description: Missing, invalid, or revoked API key
      summary: Test API key validity
      tags:
        - developer-apis/health
servers: []
...